% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mathieu.R
\name{theme_mathieu}
\alias{theme_mathieu}
\title{Mathieu's Custom Theme}
\usage{
theme_mathieu(base_size = 14, base_family = "", title.label, subtitle.label,
  x.label, y.label, x.watermark = NA, y.watermark = NA,
  watermark.size = NA)
}
\arguments{
\item{base_size}{Text size}

\item{base_family}{Text font}

\item{title.label}{Title}

\item{subtitle.label}{Subtitle}

\item{x.label}{X Label}

\item{y.label}{Y Label}

\item{x.watermark}{X coordinate for watermark}

\item{y.watermark}{Y coordinate for watermark}

\item{watermark.size}{Size of the watermark}
}
\value{
The ggplot2 theme
}
\description{
Custom ggplot2 themes for plots
}
\examples{
x.annot <- mean(player.bios$Height)
y.annot <- mean(player.bios$Weight)

ggplot(player.bios,aes(x=Position,y=Weight)) +
 geom_boxplot() +
 theme_mathieu(title.label="NHL Height by Position",
               subtitle.label="A comparison of the heights of skaters who played in the NHL",
               x.label="Position",
               y.label="Weight",
               x.watermark=2.5,
               y.watermark=y.annot,
               watermark.size=10)
}
