% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mathieu.R
\name{theme_mathieu}
\alias{theme_mathieu}
\title{Mathieu's Custom Theme}
\usage{
theme_mathieu(title.label, x.label = "", y.label = "",
  subtitle.label = "", base_size = 14, watermark.coord = NULL,
  watermark.size = 15, watermark.alpha = 0.15, watermark.angle = 30,
  watermark.color = "red")
}
\arguments{
\item{title.label}{Title}

\item{x.label}{X Label}

\item{y.label}{Y Label}

\item{subtitle.label}{Subtitle}

\item{base_size}{Font size}

\item{watermark.coord}{Coordinates for watermark}

\item{watermark.size}{Size of the watermark}

\item{watermark.alpha}{Transparency of the watermark}

\item{watermark.angle}{Angle of the watermark}

\item{watermark.color}{Color of the watermark}
}
\value{
The ggplot2 theme
}
\description{
Custom ggplot2 themes for plots
}
\examples{
ggplot(mtcars, aes(x=wt,y=qsec))
+ geom_point() +
theme_mathieu(title.label="mathieuR Theme Example",
subtitle.label="Compares wt and qsec of the mtcars dataset",
x.label = "Weight",
y.label = "QSEC",
watermark.coord=c(mean(mtcars$wt), mean(mtcars$qsec)))
}
